QBsuper sets own goal
QBSuper takes in a bool array “Dangers” and determines if there are any
dangers.  If so, calculate a vector based on which camera’s saw a
danger.  It converts that vector to a goal point to pass into AOandGtG.

the sensor locations and vector calculation are copied straight from
AOandGtG.

needs: real danger array + uses ir_array needs to be camera_array