Camera Array
So I've added the camera array.  In it's current state, it's using the
separate camera sensor file to create them, however I don't know if
they're collecting data/distances so that the boolean can be used that
will discern between an obstacle and the khepera, nor are the cones
scaling when in contact with obstacles.  Then commented out is another
version of the camera array that will scale with the obstacles because
they're just added proximity sensors, however the program won't run in
this configuration because some value that I couldn't find isn't updated
with the new number of sensors.  If you want to try running it this way
you have to update both the Quickbot.m and QBSupervisor.m files.